Playing out matches through API

+++ src/main/java/edu/bracketly/backend/model/flow/FlowHandler.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface FlowHandler {
     Match playNextMatch() throws BracketIsPlayedException;
 
-    void markAsPlayed(Long matchId, int winningSeatNumber);
+    void markAsPlayed(Long matchId, Long winningSeatNumber);
 
     BRACKET_STATUS getBracketStatus();
